updated 2d flux visualization

+++ pychron/pipeline/editors/irradiation_tray_overlay.py
@@ -37,7 +37,7 @@ class IrradiationTrayOverlay(AbstractOverlay):
         with gc:
             comp = self.component
             gc.clip_to_rect(
-                *(int(p) for p in (comp.x, comp.y, comp.width, comp.height))
+                *(float(p) for p in (comp.x, comp.y, comp.width, comp.height))
             )
             # gc.clip_to_rect(comp.x, comp.y, comp.width, comp.height)
             # gc.set_fill_color((1, 0, 1))
